get rid off polkadotProvider & set $api as global
