+++ src/components/contracts/ContractsTable.vue
@@ -58,8 +58,6 @@ export default defineComponent({
     ModalCallContract,
   },
   setup() {
-    // const { api } = useApi();
-
     const stateModal = reactive<Modal>({
       modalConfirmRemoval: false,
       modalCallContract: false,
