+++ src/components/dapp-staking/modals/ClaimRewardModal.vue
@@ -75,7 +75,6 @@ export default defineComponent({
   emits: ['update:is-open'],
   setup(props, { emit }) {
     const maxErasPerClaim = 15;
-    // const { api } = useApi();
     const store = useStore();
     const { decimal } = useChainMetadata();
     const claimInfo = ref<ClaimInfo>();
