+++ src/hooks/index.ts
@@ -1,6 +1,5 @@
 export * from './useSidebar';
 export * from './useAccount';
-// export * from './useApi';
 export * from './useBalance';
 export * from './useContracts';
 export * from './useMessages';
